chore: remove modules that only contained tests
Due to refactoring the issue tracking into a single `track_issue` function, these modules didn't end
up having any code, just tests.

The tests were moved to separate testing modules in the `detection` module.
