+++ sn_dysfunction/src/detection.rs
@@ -352,3 +352,429 @@ mod tests {
         }
     }
 }
+
+#[cfg(test)]
+mod ops_tests {
+    use crate::tests::init_test_logger;
+    use crate::{error::Result, DysfunctionDetection, DysfunctionSeverity, IssueType, OperationId};
+    use rand::Rng;
+    use xor_name::{rand::random as random_xorname, XorName};
+
+    // we can see 500 pending issues under load
+    pub(crate) const NORMAL_OPERATIONS_ISSUES: usize = 500;
+    // Above this, nodes should be sus
+    pub(crate) const SUSPECT_OPERATIONS_ISSUES: usize = 800;
+    pub(crate) const DYSFUNCTIONAL_OPERATIONS_ISSUES: usize = 1500;
+
+    fn get_random_operation_id() -> Option<OperationId> {
+        let mut rng = rand::thread_rng();
+        Some(rng.gen())
+    }
+
+    #[tokio::test]
+    async fn op_dysfunction_no_variance_is_okay() -> Result<()> {
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+        for adult in &adults {
+            for _ in 0..NORMAL_OPERATIONS_ISSUES {
+                let _ = dysfunctional_detection
+                    .track_issue(
+                        *adult,
+                        IssueType::PendingRequestOperation(get_random_operation_id()),
+                    )
+                    .await;
+            }
+        }
+
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+                .await?
+                .len(),
+            0
+        );
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+                .await?
+                .len(),
+            0
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn op_dysfunction_basics() -> Result<()> {
+        init_test_logger();
+        let _outer_span = tracing::info_span!("op_dysfunction_basics").entered();
+
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        for adult in &adults {
+            for _ in 0..NORMAL_OPERATIONS_ISSUES {
+                let _ = dysfunctional_detection
+                    .track_issue(
+                        *adult,
+                        IssueType::PendingRequestOperation(get_random_operation_id()),
+                    )
+                    .await;
+            }
+        }
+
+        let new_adult = random_xorname();
+        let _ = dysfunctional_detection.add_new_node(new_adult).await;
+        for _ in 0..SUSPECT_OPERATIONS_ISSUES {
+            let _ = dysfunctional_detection
+                .track_issue(
+                    new_adult,
+                    IssueType::PendingRequestOperation(get_random_operation_id()),
+                )
+                .await;
+        }
+
+        assert!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+                .await?
+                .contains(&new_adult),
+            "node is not showing as sus"
+        );
+
+        assert!(
+            !dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+                .await?
+                .contains(&new_adult),
+            "node is dysfunctional when it should only be sus"
+        );
+
+        for _ in 0..DYSFUNCTIONAL_OPERATIONS_ISSUES - SUSPECT_OPERATIONS_ISSUES {
+            let _ = dysfunctional_detection
+                .track_issue(
+                    new_adult,
+                    IssueType::PendingRequestOperation(get_random_operation_id()),
+                )
+                .await;
+        }
+
+        assert!(dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?
+            .contains(&new_adult));
+
+        assert!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+                .await?
+                .contains(&new_adult),
+            "our node has not been listed as dysfunctional"
+        );
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod comm_tests {
+    use crate::tests::init_test_logger;
+    use crate::{DysfunctionDetection, DysfunctionSeverity, IssueType};
+
+    use eyre::Error;
+    use xor_name::{rand::random as random_xorname, XorName};
+
+    type Result<T, E = Error> = std::result::Result<T, E>;
+
+    // Above this, nodes should be sus
+    // this is only counting last RECENT minutes atm
+    pub(crate) const NORMAL_CONNECTION_PROBLEM_COUNT: usize = 5;
+    pub(crate) const SUSPECT_CONNECTION_PROBLEM_COUNT: usize = 20;
+    pub(crate) const DYSFUNCTIONAL_CONNECTION_PROBLEM_COUNT: usize = 35;
+
+    #[tokio::test]
+    async fn conn_dys_is_tolerant_of_norms() -> Result<()> {
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        for adult in &adults {
+            for _ in 0..NORMAL_CONNECTION_PROBLEM_COUNT {
+                dysfunctional_detection
+                    .track_issue(*adult, IssueType::Communication)
+                    .await?;
+            }
+        }
+
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+                .await?
+                .len(),
+            0,
+            "no nodes are dysfunctional"
+        );
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+                .await?
+                .len(),
+            0,
+            "no nodes are suspect"
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn conn_dysfunction_basics_sus_comes_first() -> Result<()> {
+        init_test_logger();
+        let _outer_span = tracing::info_span!("conn_dysfunction_basics_sus_comes_first").entered();
+
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        for adult in &adults {
+            for _ in 0..NORMAL_CONNECTION_PROBLEM_COUNT {
+                dysfunctional_detection
+                    .track_issue(*adult, IssueType::Communication)
+                    .await?;
+            }
+        }
+
+        let new_adult = random_xorname();
+        let _ = dysfunctional_detection.add_new_node(new_adult).await;
+        for _ in 0..SUSPECT_CONNECTION_PROBLEM_COUNT {
+            dysfunctional_detection
+                .track_issue(new_adult, IssueType::Communication)
+                .await?;
+        }
+
+        let sus = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?;
+        assert_eq!(sus.len(), 1, "only one node is sus");
+        assert!(sus.contains(&new_adult), "our adult is not sus");
+
+        let dysfunctional_nodes = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+            .await?;
+
+        assert!(
+            !dysfunctional_nodes.contains(&new_adult),
+            "our added node is dysfunctional when it should not be"
+        );
+
+        assert_eq!(
+            dysfunctional_nodes.len(),
+            0,
+            "more nodes are dysfunctional than they should be"
+        );
+
+        for _ in 0..DYSFUNCTIONAL_CONNECTION_PROBLEM_COUNT - SUSPECT_CONNECTION_PROBLEM_COUNT {
+            dysfunctional_detection
+                .track_issue(new_adult, IssueType::Communication)
+                .await?;
+        }
+
+        let sus = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?;
+        assert!(sus.contains(&new_adult), "our adult is still sus");
+        assert_eq!(sus.len(), 1, "only one adult is sus");
+
+        debug!("=============================");
+        let dysfunctional_nodes = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+            .await?;
+
+        assert!(
+            dysfunctional_nodes.contains(&new_adult),
+            "our adult should now be dysfunctional but is not"
+        );
+        assert_eq!(
+            dysfunctional_nodes.len(),
+            1,
+            "our node should be the only dysfunctional nodes"
+        );
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod knowledge_tests {
+    use crate::tests::init_test_logger;
+    use crate::{DysfunctionDetection, DysfunctionSeverity, IssueType};
+
+    use eyre::Error;
+    use xor_name::{rand::random as random_xorname, XorName};
+
+    type Result<T, E = Error> = std::result::Result<T, E>;
+
+    // 5 here means we have some tolerance for AE rounds while nodes are getting up to speed on churn/split
+    pub(crate) const NORMAL_KNOWLEDGE_ISSUES: usize = 7;
+    pub(crate) const SUSPECT_KNOWLEDGE_ISSUES: usize = 15;
+    pub(crate) const DYSFUNCTIONAL_KNOWLEDGE_ISSUES: usize = 28;
+
+    #[tokio::test]
+    async fn knowledge_dys_is_tolerant_of_norms() -> Result<()> {
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        // Write data NORMAL_KNOWLEDGE_ISSUES times to the 10 adults
+        for adult in &adults {
+            for _ in 0..NORMAL_KNOWLEDGE_ISSUES {
+                dysfunctional_detection
+                    .track_issue(*adult, IssueType::Knowledge)
+                    .await?;
+            }
+        }
+
+        // Assert there are not any dysfuncitonal nodes
+        // This is because all of them are within the tolerance ratio of each other
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+                .await?
+                .len(),
+            0,
+            "no nodes are dysfunctional"
+        );
+        assert_eq!(
+            dysfunctional_detection
+                .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+                .await?
+                .len(),
+            0,
+            "no nodes are suspect"
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn knowledge_dys_is_not_too_sharp() -> Result<()> {
+        init_test_logger();
+        let _outer_span = tracing::info_span!("knowledge_dys_is_not_too_sharp").entered();
+
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        // Add a new adults
+        let new_adult = random_xorname();
+        let _ = dysfunctional_detection.add_new_node(new_adult).await;
+
+        // Add just one knowledge issue...
+        for _ in 0..1 {
+            dysfunctional_detection
+                .track_issue(new_adult, IssueType::Knowledge)
+                .await?;
+        }
+
+        let sus = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?;
+
+        // Assert that the new adult is not detected as suspect.
+        assert!(!sus.contains(&new_adult), "our adult should not be sus");
+        assert_eq!(sus.len(), 0, "no node is sus");
+
+        let dysfunctional_nodes = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+            .await?;
+
+        // Assert that the new adult is not dysfuncitonal
+        assert!(
+            !dysfunctional_nodes.contains(&new_adult),
+            "our adult should not be dysfunctional"
+        );
+        assert_eq!(
+            dysfunctional_nodes.len(),
+            0,
+            "no node is dysfunctional node"
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn knowledge_dysfunction_basics_sus_comes_first() -> Result<()> {
+        init_test_logger();
+        let _outer_span =
+            tracing::info_span!("knowledge_dysfunction_basics_sus_comes_first").entered();
+
+        let adults = (0..10).map(|_| random_xorname()).collect::<Vec<XorName>>();
+
+        let dysfunctional_detection = DysfunctionDetection::new(adults.clone());
+
+        for adult in &adults {
+            for _ in 0..NORMAL_KNOWLEDGE_ISSUES {
+                dysfunctional_detection
+                    .track_issue(*adult, IssueType::Knowledge)
+                    .await?;
+            }
+        }
+
+        let new_adult = random_xorname();
+        let _ = dysfunctional_detection.add_new_node(new_adult).await;
+        for _ in 0..SUSPECT_KNOWLEDGE_ISSUES {
+            dysfunctional_detection
+                .track_issue(new_adult, IssueType::Knowledge)
+                .await?;
+        }
+
+        let sus = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?;
+        assert_eq!(sus.len(), 1, "only one node is sus");
+        assert!(sus.contains(&new_adult), "our adult is not sus");
+
+        let dysfunctional_nodes = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+            .await?;
+
+        assert!(
+            !dysfunctional_nodes.contains(&new_adult),
+            "our added node is dysfunctional when it should not be"
+        );
+
+        assert_eq!(
+            dysfunctional_nodes.len(),
+            0,
+            "more nodes are dysfunctional than they should be"
+        );
+
+        for _ in 0..DYSFUNCTIONAL_KNOWLEDGE_ISSUES - SUSPECT_KNOWLEDGE_ISSUES {
+            dysfunctional_detection
+                .track_issue(new_adult, IssueType::Knowledge)
+                .await?;
+        }
+
+        let sus = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Suspicious)
+            .await?;
+        assert!(sus.contains(&new_adult), "our adult is still sus");
+        assert_eq!(sus.len(), 1, "only one adult is sus");
+
+        let dysfunctional_nodes = dysfunctional_detection
+            .get_nodes_beyond_severity(DysfunctionSeverity::Dysfunctional)
+            .await?;
+
+        assert!(
+            dysfunctional_nodes.contains(&new_adult),
+            "our adult should now be dysfunctional, but is not"
+        );
+        assert_eq!(
+            dysfunctional_nodes.len(),
+            1,
+            "our adult is the only dysfunctional node"
+        );
+
+        Ok(())
+    }
+}
